logout bug fixed while admin was cliking

+++ src/pages/Profile/Profile.jsx
@@ -58,12 +58,15 @@ const Profile = () => {
   }, [id])
 
   const logout = () => {
-    localStorage.clear()
-    dispatch({
-      type: actionTypes.LOGOUT,
-      payload: {},
-    })
-    navigate("/")
+    if (id == authentication.userId) {
+      localStorage.clear()
+      dispatch({
+        type: actionTypes.LOGOUT,
+        payload: {},
+      })
+      navigate("/")
+    }
+
   }
 
   const removeStudent = () => {
